requirements.txt
django==4.2.5
scikit-learn==1.0.2
pandas== 2.1.1
djangorestframework

